# internal/http/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for the BengoBox notifications service.",
        "title": "Notifications Service API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/healthz": {
            "get": {
                "description": "Returns OK when the API process is running.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Service liveness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.livenessResponse"
                        }
                    }
                }
            }
        },
        "/metrics": {
            "get": {
                "description": "Exposes Prometheus metrics for scraping.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Prometheus metrics",
                "responses": {
                    "200": {
                        "description": "Prometheus metrics payload",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/readyz": {
            "get": {
                "description": "Validates connectivity to Postgres, Redis, and NATS (when configured).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Readiness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.readinessResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.readinessResponse"
                        }
                    }
                }
            }
        },
        "/v1/{tenantId}/notifications/messages": {
            "post": {
                "description": "Accepts a notification payload and queues it for downstream processing.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Queue notification message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant identifier",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/handlers.enqueueResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/v1/{tenantId}/templates": {
            "get": {
                "description": "Returns the default set of notification templates available to the tenant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Templates"
                ],
                "summary": "List notification templates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant identifier",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.templateListResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.CreateMessageRequest": {
            "type": "object",
            "required": [
                "channel",
                "data",
                "template",
                "tenant",
                "to"
            ],
            "properties": {
                "channel": {
                    "type": "string"
                },
                "data": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "template": {
                    "type": "string"
                },
                "tenant": {
                    "type": "string"
                },
                "to": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "handlers.enqueueResponse": {
            "type": "object",
            "properties": {
                "requestId": {
                    "type": "string",
                    "example": "req_123"
                },
                "status": {
                    "type": "string",
                    "example": "queued"
                }
            }
        },
        "handlers.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation failed"
                }
            }
        },
        "handlers.livenessResponse": {
            "type": "object",
            "properties": {
                "service": {
                    "type": "string",
                    "example": "notifications-app"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "handlers.readinessResponse": {
            "type": "object",
            "properties": {
                "dependencies": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "handlers.templateListResponse": {
            "type": "object",
            "properties": {
                "templates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.templateSummary"
                    }
                }
            }
        },
        "handlers.templateSummary": {
            "type": "object",
            "properties": {
                "channel": {
                    "type": "string",
                    "example": "email"
                },
                "id": {
                    "type": "string",
                    "example": "payment_success"
                }
            }
        }
    }
}